use raw_input instead of input for security reasons
also fixes your parsing problems with spaces.
input() actually runs whatever is entered by the user
as Python code.
When it was still input(), you could type something like
[board.__setitem__(3, [1,1,1,1,1,1,1]),5][-1]
in as player 1 and win on the first move.
raw_input() always returns a string, which is safer

the other changes aren't so important